move type helper classes to seperate file
